pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "formula"
version = "0.0.1"
description = "Workbench module to run Excel or Python formulas"
authors = ["Adam Hooper <adam@adamhooper.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
cjwmodule = "~=4.1"
cjwpandasmodule = "~=0.2.0"
formulas = "~=1.0.0"
pyarrow = "~=4.0"
python = "~=3.8.0"

[tool.poetry.dev-dependencies]
pytest = "~=6.0"
pytest-cov = "~= 2.10"

[tool.isort]
# Black compatibility
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = True
skipsdist = True
envlist = py38-{pyflakes,black,isort,pytest}

[flake8]
exclude = venv/*,tox/*,specs/*,build/*
ignore = E123,E128,E266,E402,W503,E731,W601
max-line-length = 88

[testenv]
deps =
  pyflakes: pyflakes>=2.2
  black: black
  isort: isort
  pytest: poetry
commands_pre =
  pytest: poetry install -v
commands =
  pyflakes: pyflakes formula.py tests
  black: black --check formula.py tests
  isort: isort --check --diff formula.py tests
  pytest: poetry run pytest --cov=formula --cov-report term-missing -v
"""
